requirements.txt
langchain
openai
fastapi
uvicorn
python-dotenv
streamlit
requests
langchain-community
langchain_classic
